One Altera platform build issue, and a number of minor sync-ups with "develop"
components/file_read.rcc/file_read.c: merge Aaron's MIF EOF fix, and add some others
components/file_read_msg.rcc/file_read_msg.c: tweaks from develop
components/file_write_msg.rcc/file_write_msg.c: ditto
hdl/devices/lime_rx_proxy.rcc/lime_shared.h: ditto
os/interfaces/include/OcpiOsDataTypes.h: ditto
os/interfaces/include/OcpiOsEther.h: ditto
os/interfaces/include/OcpiOsStdint.h: ditto
os/linux/driver/opencpi.c: ditto
os/linux/src/OcpiOsMisc.cxx: ditto
os/linux/src/OcpiOsServerSocket.cxx: ditto
os/linux/src/OcpiOsSocket.cxx: ditto
runtime/application/include/OcpiApi.h: ditto
runtime/ctests/src/UtGenericLoopBackWorkers.c: ditto
runtime/hdl-support/src/ocpinft_main.c: ditto
runtime/hdl/include/HdlAccess.h: ditto
runtime/hdl/include/HdlDevice.h: ditto
runtime/util/misc/include/OcpiUtilMisc.h: ditto
runtime/util/property/include/OcpiUtilDataTypesApi.h: ditto

tools/cdk/include/hdl/hdl-assembly.mk: improved message - still not the best
tools/cdk/include/hdl/hdl-platform.mk: Fix altera platform build case
runtime/hdl-support/src/HdlSimServer.cxx: bug fix with spin timer not getting started
-- caused occasional assertion failed...
examples/dds_ports/standalone_dds/Test/C++/src/CheckStatus.h: strip CRs
examples/dds_ports/standalone_dds/Test/C++/src/DDSEntityManager.h: dito

+++ hdl/devices/lime_rx_proxy.rcc/lime_shared.h
@@ -7,7 +7,7 @@ namespace OCPI {
   struct Divider {
     uint8_t nint, nfrac_hi, nfrac_mid, nfrac_lo;
   };
-  void calcDividers(float pll_hz, float lo_hz, Divider &div);
+  void calcDividers(double pll_hz, double lo_hz, Divider &div);
   // Return a string error if there was one.  Otherwise NULL
   const char
     *setVcoCap(uint8_t (*readVtune)(void *arg), void (*writeVcoCap)(void *arg, uint8_t val),
